readme C++
I had to practice editing a branch to get a hang of Github , as described so precisely by the guide. So, I tried and succeeded in making a change. As you might see, I went on to write the things I wanted to know in C++, just the simple hello world code, me messing around with it. I never intend to learn C++ soon. But C++ does interest me.